requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.6
django-crispy-forms==2.0
Pillow==10.0.1
psycopg2==2.9.9
python-decouple==3.8
sqlparse==0.4.4
tzdata==2023.3
uuid==1.30
